pyproject.toml
[tool.poetry]
name = "granton-test-assignment"
version = "0.1.0"
description = ""
authors = ["MK-444 <likefi444@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
pre-commit = "^3.7.1"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"
g4f = "^0.3.1.9"
pytest-asyncio = "^0.23.7"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
flake8 = "^7.0.0"
pylint = "^3.2.2"
black = "^24.4.2"
ruff = "^0.4.5"
mypy = "^1.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
